reorg and move out the C API
